Comienza modularización: se agregan componentes Intro.js y GameDisplay.js. Se corrige el directorio de la data de cartas howl. Algunos cambios en CSS.
