+++ src/data/howl/howl.js
@@ -1,15 +1,15 @@
 export default {
     name: 'Howl',
     items: [
-      { id: 'calucipher', image: 'src/data/howl/images/calcipher.png', bgColor: '#ffffff' },
-      { id: 'calcipher2', image: '/images/calcipher2.png', bgColor: '#ffffff' },
-      { id: 'castle', image: '/images/castle.png', bgColor: '#ffffff' },
-      { id: 'dog', image: '/images/dog.png', bgColor: '#ffffff' },
-      { id: 'howl', image: '/images/howl.png', bgColor: '#ffffff' },
-      { id: 'markl', image: '/images/markl.png', bgColor: '#ffffff' },
-      { id: 'sophie', image: '/images/sophie.png', bgColor: '#ffffff' },
-      { id: 'thewitch', image: '/images/thewitch.png', bgColor: '#ffffff' },
-      { id: 'turniphead', image: '/images/turniphead.png', bgColor: '#ffffff' },
+      { id: 'calucipher', image: 'data/howl/images/calcipher.png', bgColor: '#ffffff' },
+      { id: 'calcipher2', image: 'data/howl/images/calcipher2.png', bgColor: '#ffffff' },
+      { id: 'castle', image: 'data/howl/images/castle.png', bgColor: '#ffffff' },
+      { id: 'dog', image: 'data/howl/images/dog.png', bgColor: '#ffffff' },
+      { id: 'howl', image: 'data/howl/images/howl.png', bgColor: '#ffffff' },
+      { id: 'markl', image: 'data/howl/images/markl.png', bgColor: '#ffffff' },
+      { id: 'sophie', image: 'data/howl/images/sophie.png', bgColor: '#ffffff' },
+      { id: 'thewitch', image: 'data/howl/images/thewitch.png', bgColor: '#ffffff' },
+      { id: 'turniphead', image: 'data/howl/images/turniphead.png', bgColor: '#ffffff' },
     ],
   };
