+++ src/index.html
@@ -4,15 +4,11 @@
 <head>
   <meta charset="utf-8">
   <title>Memory Match</title>
-  <link rel="preconnect" href="https://fonts.gstatic.com">
-  <link href="https://fonts.googleapis.com/css2?family=Orbitron:wght@800&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="style.css" />
 </head>
 
 <body>
   <div id="root">
-    <!-- <button id="play" class="playButton">JUGAR!</button> -->
-  </div>
   </div>
   <script src="main.js" type="module"></script>
   </main>
